feat(WELCOME): add open graph metadata
http://ogp.me/
The Open Graph protocol enables any web page to become a rich object in
a social graph. For instance, this is used on Facebook to allow any web
page to have the same functionality as any other object on Facebook.

Hack was implemented to prevent problems from hashes added by webpack
to filenames. Hash is not predictable simply, so choice was to copy
image directly from assets to /dist without transformation. og:metadata
can, by this way, refer to a project file. Assumed that these files are
stable.